Refactor compatibility package API
