--- src/pytest_bdd/compatibility/__init__.py
@@ -1,6 +1,0 @@
-"""Common type definitions."""
-from __future__ import annotations
-
-from typing import Literal, Protocol, runtime_checkable
-
-__all__ = ["Literal", "Protocol", "runtime_checkable"]
