+++ src/pytest_bdd/compatibility/parser.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from pathlib import Path
 from typing import TYPE_CHECKING, Callable
 
-from pytest_bdd.compatibility import Protocol, runtime_checkable
 from pytest_bdd.compatibility.pytest import Config
+from pytest_bdd.compatibility.typing import Protocol, runtime_checkable
 from pytest_bdd.utils import PytestBDDIdGeneratorHandler
 
 if TYPE_CHECKING:  # pragma: no cover
