+++ src/pytest_bdd/cucumber_json.py
@@ -7,8 +7,8 @@ import os
 import time
 from typing import TYPE_CHECKING, Any, cast
 
-from pytest_bdd.compatibility import Protocol, runtime_checkable
 from pytest_bdd.compatibility.pytest import Parser, TerminalReporter, TestReport
+from pytest_bdd.compatibility.typing import Protocol, runtime_checkable
 
 if TYPE_CHECKING:  # pragma: no cover
     from pytest_bdd.compatibility.pytest import Config as BaseConfig
